Sort of allow changing the user-backend.
The changes make it to some extend possible to promote a user to another
backend, e.g. from Database -> LDAP. This is not a real transfer ATM,
i.e. the original user is not deleled. It is hence also one-way as an
already existing user with the same uid in the new backend blocks the
transfer.