Storage\Common::isDeleteable() -- dirname() returns . for top-level node.
dirname() is in principle not wrong here, however, the convention seems
to be that the paths passed here look relative but are in fact only
relative to the mount-point. This means a mount

MOUNTPOINT/foobar

with an in fact deletable file will end up here with the path foobar
WITHOUT a leading slash or any top-level directory. So dirname() will
return . and not ''.