Files\Node\HookConnector: fix node type for non-existing rename and copy sources/targets.
Currently, HookConnector::getNodeForPath() just asks
Filesystem::is_dir($path) if a FileInfo for $path cannot be found.
However, if $path is a non-existing target in the pre-rename or pre-copy
events, or the then no-longer existing source in the post-rename event
this will fail: Filesystem::is_dir($path) will return false (the $path
does not event exist), and the a NonExistingFile node is generated in
any case.

This is at least surprising: when renaming an existing folder to
non-existing target I really would expect a NonExistingFolder node as
target in the pre-rename event, and the same for the source in the
post-rename event.