private/User/Manager.php: make the ordering of backends predictable.
The UserManager maintains several backends, but does not maintain a
predictable ordering of backends, so it can happen that createUser uses
one ordering and then other methods use another ordering (createUser()
in particular uses the database-backend last, so at least there is a
source of a conflict.

The commit also changes createUserFromBackend(BACKEND) to only check if the
user exists in that given backend.