DB: only use the PrimaryReadReplica connection if there is a replication database.
Why? The PrimaryReadReplica connection is more complicated and thus --
in the absence of a replication databse -- make things more complicated
and error prone without boosting the performance. This commit therefore
only uses the more complicated primary-read-replica connection if that
actually is a replication database and otherwise falls back to the
simple standard connection.