NC-PR: apps/dav fix json_decode to StdObject and fix resource access restrictions.
json_decode in this context must decode to associative array.

More important: the group_restrictions column contains JSON data
referring to the auto-increment index of the calresources_restricts
table. Consequently it does not make sense to compare -- even not after
JSON decode in the correct format -- directly to the group-membership of
the active user. We first need another query to the
calresources_restricts table which translates the AI index values to
NC group-ids.